#80 [REFACTOR] 이미지 유틸 클래스 패키지 이전
