+++ src/main/java/com/wakeUpTogetUp/togetUp/api/mission/service/MissionImageService.java
@@ -2,7 +2,7 @@ package com.wakeUpTogetUp.togetUp.api.mission.service;
 
 import com.wakeUpTogetUp.togetUp.api.mission.domain.CustomAnalysisEntity;
 import com.wakeUpTogetUp.togetUp.utils.FileUtil;
-import com.wakeUpTogetUp.togetUp.utils.ImageDrawer;
+import com.wakeUpTogetUp.togetUp.utils.image.ImageDrawer;
 import com.wakeUpTogetUp.togetUp.infra.aws.s3.model.CustomFile;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
