+++ src/main/java/com/wakeUpTogetUp/togetUp/utils/image/ImageUtil.java
@@ -1,9 +1,10 @@
-package com.wakeUpTogetUp.togetUp.utils;
+package com.wakeUpTogetUp.togetUp.utils.image;
 
 import static org.apache.commons.imaging.Imaging.getMetadata;
 
 import com.wakeUpTogetUp.togetUp.common.Status;
 import com.wakeUpTogetUp.togetUp.exception.BaseException;
+import com.wakeUpTogetUp.togetUp.utils.FileUtil;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
